refactor account store and emit login and logout events via bus
